kaping/requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
numpy==1.26.1
ReFinED @ https://github.com/amazon-science/ReFinED/archive/refs/tags/V1.zip
scikit-learn==1.3.2
sentence-transformers==2.2.2
transformers==4.34.1
requests
torch>=1.12.1
nltk>=3.5
urllib3>=1.25.8
Unidecode>=1.1.1
ujson
tqdm
tensorboard
Cython
lmdb>=1.0
boto3
prettyprint
botocore
google-generativeai